% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\docType{data}
\name{concept}
\alias{concept}
\title{concept}
\format{
A data frame with 2355 rows and 10 variables:
\describe{
\item{concept_name}{\emph{character}}
\item{domain_id}{\emph{character}}
\item{vocabulary_id}{\emph{character}}
\item{concept_class_id}{\emph{character}}
\item{standard_concept}{\emph{character}}
\item{concept_code}{\emph{character}}
\item{valid_start_date}{\emph{Date}}
\item{valid_end_date}{\emph{Date}}
\item{invalid_reason}{\emph{character}}
\item{concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#concept }
}
\usage{
concept
}
\description{

}
\keyword{datasets}
