% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_class.R
\docType{data}
\name{concept_class}
\alias{concept_class}
\title{concept_class}
\format{
A data frame with 0 rows and 3 variables:
\describe{
\item{concept_class_id}{A unique key for each class. }
\item{concept_class_name}{The name describing the Concept Class, e.g. Clinical Finding, Ingredient, etc. }
\item{concept_class_concept_id}{A Concept that represents the Concept Class. }
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#concept_class}
}
\usage{
concept_class
}
\description{
The CONCEPT_CLASS table is a reference table, which includes a list of the classifications used to differentiate Concepts within a given Vocabulary. This reference table is populated with a single record for each Concept Class.
}
\keyword{datasets}
