% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_synonym.R
\docType{data}
\name{concept_synonym}
\alias{concept_synonym}
\title{concept_synonym}
\format{
A data frame with XX rows and XX variables:
\describe{
\item{concept_id }{}
\item{concept_synonym_name}{}
\item{language_concept_id}{}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#concept_synonym}
}
\usage{
concept_synonym
}
\description{
\subsection{Table Description}{

The CONCEPT_SYNONYM table is used to store alternate names and descriptions for Concepts.
}
}
\keyword{datasets}
