% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_era.R
\docType{data}
\name{condition_era}
\alias{condition_era}
\title{condition_era}
\format{
A data frame with 688 rows and 6 variables:
\describe{
\item{condition_era_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{condition_concept_id}{\emph{integer}}
\item{condition_era_start_date}{\emph{Date}}
\item{condition_era_end_date}{\emph{Date}}
\item{condition_occurrence_count}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#condition_era }
}
\usage{
condition_era
}
\description{

}
\keyword{datasets}
