% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\docType{data}
\name{death}
\alias{death}
\title{death}
\format{
A data frame with 2 rows and 7 variables:
\describe{
\item{person_id}{\emph{integer}}
\item{death_date}{\emph{Date}}
\item{death_datetime}{\emph{POSIXct}}
\item{death_type_concept_id}{\emph{integer}}
\item{cause_concept_id}{\emph{integer}}
\item{cause_source_value}{\emph{character}}
\item{cause_source_concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#death }
}
\usage{
death
}
\description{

}
\keyword{datasets}
