% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\docType{data}
\name{domain}
\alias{domain}
\title{domain}
\format{
A data frame with 0 rows and 3 variables:
\describe{
\item{domain_id}{\emph{character}}
\item{domain_name}{\emph{character}}
\item{domain_concept_id}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#domain }
}
\usage{
domain
}
\description{

}
\keyword{datasets}
