% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_era.R
\docType{data}
\name{drug_era}
\alias{drug_era}
\title{drug_era}
\format{
A data frame with 344 rows and 7 variables:
\describe{
\item{drug_era_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{drug_concept_id}{\emph{integer}}
\item{drug_era_start_date}{\emph{Date}}
\item{drug_era_end_date}{\emph{Date}}
\item{drug_exposure_count}{\emph{integer}}
\item{gap_days}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_era }
}
\usage{
drug_era
}
\description{

}
\keyword{datasets}
