% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_strength.R
\docType{data}
\name{drug_strength}
\alias{drug_strength}
\title{drug_strength}
\format{
A data frame with 475035 rows and 9 variables:
\describe{
\item{amount_value}{\emph{numeric}}
\item{numerator_value}{\emph{numeric}}
\item{denominator_value}{\emph{numeric}}
\item{box_size}{\emph{integer}}
\item{valid_start_date}{\emph{Date}}
\item{valid_end_date}{\emph{Date}}
\item{invalid_reason}{\emph{character}}
\item{ingredient_concept_id}{\emph{integer}}
\item{drug_concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_strength }
}
\usage{
drug_strength
}
\description{

}
\keyword{datasets}
