% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.R
\docType{data}
\name{measurement}
\alias{measurement}
\title{measurement}
\format{
A data frame with 194 rows and 18 variables:
\describe{
\item{measurement_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{measurement_concept_id}{\emph{integer}}
\item{measurement_date}{\emph{Date}}
\item{measurement_datetime}{\emph{POSIXct}}
\item{measurement_type_concept_id}{\emph{integer}}
\item{operator_concept_id}{\emph{integer}}
\item{value_as_number}{\emph{numeric}}
\item{value_as_concept_id}{\emph{integer}}
\item{unit_concept_id}{\emph{integer}}
\item{range_low}{\emph{numeric}}
\item{range_high}{\emph{numeric}}
\item{provider_id}{\emph{integer}}
\item{visit_occurrence_id}{\emph{integer}}
\item{measurement_source_value}{\emph{character}}
\item{measurement_source_concept_id}{\emph{integer}}
\item{unit_source_value}{\emph{character}}
\item{value_source_value}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#measurement }
}
\usage{
measurement
}
\description{

}
\keyword{datasets}
