% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_period.R
\docType{data}
\name{observation_period}
\alias{observation_period}
\title{observation_period}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{observation_period_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{observation_period_start_date}{\emph{Date}}
\item{observation_period_end_date}{\emph{Date}}
\item{period_type_concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#observation_period }
}
\usage{
observation_period
}
\description{

}
\keyword{datasets}
