% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payer_plan_period.R
\docType{data}
\name{payer_plan_period}
\alias{payer_plan_period}
\title{payer_plan_period}
\format{
A data frame with 33 rows and 7 variables:
\describe{
\item{payer_plan_period_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{payer_plan_period_start_date}{\emph{Date}}
\item{payer_plan_period_end_date}{\emph{Date}}
\item{payer_source_value}{\emph{character}}
\item{plan_source_value}{\emph{character}}
\item{family_source_value}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#payer_plan_period }
}
\usage{
payer_plan_period
}
\description{

}
\keyword{datasets}
