% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\docType{data}
\name{person}
\alias{person}
\title{person}
\format{
A data frame with 10 rows and 18 variables:
\describe{
\item{person_id}{\emph{integer}}
\item{gender_concept_id}{\emph{integer}}
\item{year_of_birth}{\emph{integer}}
\item{month_of_birth}{\emph{integer}}
\item{day_of_birth}{\emph{integer}}
\item{birth_datetime}{\emph{POSIXct}}
\item{race_concept_id}{\emph{integer}}
\item{ethnicity_concept_id}{\emph{integer}}
\item{location_id}{\emph{integer}}
\item{provider_id}{\emph{integer}}
\item{care_site_id}{\emph{integer}}
\item{person_source_value}{\emph{character}}
\item{gender_source_value}{\emph{character}}
\item{gender_source_concept_id}{\emph{integer}}
\item{race_source_value}{\emph{character}}
\item{race_source_concept_id}{\emph{integer}}
\item{ethnicity_source_value}{\emph{character}}
\item{ethnicity_source_concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#person }
}
\usage{
person
}
\description{

}
\keyword{datasets}
