% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\docType{data}
\name{relationship}
\alias{relationship}
\title{relationship}
\format{
A data frame with 0 rows and 6 variables:
\describe{
\item{relationship_id}{\emph{character}}
\item{relationship_name}{\emph{character}}
\item{is_hierarchical}{\emph{character}}
\item{defines_ancestry}{\emph{character}}
\item{reverse_relationship_id}{\emph{character}}
\item{relationship_concept_id}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#relationship }
}
\usage{
relationship
}
\description{

}
\keyword{datasets}
