% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synPUF.R
\docType{package}
\name{synPUF}
\alias{synPUF}
\title{synPUF: 50 Person Subset of DE-SynPUF}
\description{
This package contains a 50 person subset of the CMS 2008-2010 Data Entrepreneurs’
Synthetic Public Use File (DE-SynPUF) from OHDSI.
}
