% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\docType{data}
\name{vocabulary}
\alias{vocabulary}
\title{vocabulary}
\format{
A data frame with 0 rows and 5 variables:
\describe{
\item{vocabulary_id}{\emph{character}}
\item{vocabulary_name}{\emph{character}}
\item{vocabulary_reference}{\emph{character}}
\item{vocabulary_version}{\emph{character}}
\item{vocabulary_concept_id}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#vocabulary }
}
\usage{
vocabulary
}
\description{

}
\keyword{datasets}
