% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_site.R
\docType{data}
\name{care_site}
\alias{care_site}
\title{care_site}
\format{
A data frame with 0 rows and 6 variables:
\describe{
\item{care_site_id}{\emph{character}}
\item{care_site_name}{\emph{character}}
\item{place_of_service_concept_id}{\emph{character}}
\item{location_id}{\emph{character}}
\item{care_site_source_value}{\emph{character}}
\item{place_of_service_source_value}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#care_site }
}
\usage{
care_site
}
\description{

}
\keyword{datasets}
