% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_site.R
\docType{data}
\name{care_site}
\alias{care_site}
\title{care_site}
\format{
A data frame with 0 rows and 6 variables:
\describe{
\item{care_site_id}{A unique identifier for each Care Site.}
\item{care_site_name}{The verbatim description or name of the Care Site as in data source}
\item{place_of_service_concept_id}{This is a high-level way of characterizing a Care Site. Typically, however, Care Sites can provide care in multiple settings (inpatient, outpatient, etc.) and this granularity should be reflected in the visit. }
\item{location_id}{A foreign key that refers to a Place of Service Concept ID in the Standardized Vocabularies.}
\item{care_site_source_value}{The identifier for the Care Site in the source data, stored here for reference.}
\item{place_of_service_source_value}{The source code for the Place of Service as it appears in the source data, stored here for reference.}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#care_site}
}
\usage{
care_site
}
\description{
The CARE_SITE table contains a list of uniquely identified institutional (physical or organizational) units where healthcare delivery is practiced (offices, wards, hospitals, clinics, etc.).
}
\keyword{datasets}
