% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_era.R
\docType{data}
\name{condition_era}
\alias{condition_era}
\title{condition_era}
\format{
A data frame with 3465 rows and 6 variables:
\describe{
\item{condition_era_id}{}
\item{person_id}{}
\item{condition_concept_id}{The Concept Id representing the Condition. }
\item{condition_era_start_date}{The start date for the Condition Era constructed from the individual instances of Condition Occurrences. It is the start date of the very first chronologically recorded instance of the condition with at least 31 days since any prior record of the same Condition. }
\item{condition_era_end_date }{The end date for the Condition Era constructed from the individual instances of Condition Occurrences. It is the end date of the final continuously recorded instance of the Condition. }
\item{condition_occurrence_count}{t 	The number of individual Condition Occurrences used to construct the condition era. }
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#condition_era}
}
\usage{
condition_era
}
\description{
\subsection{Table Description}{

A Condition Era is defined as a span of time when the Person is assumed to have a given condition. Similar to Drug Eras, Condition Eras are chronological periods of Condition Occurrence. Combining individual Condition Occurrences into a single Condition Era serves two purposes:
\itemize{
\item{It allows aggregation of chronic conditions that require frequent ongoing care, instead of treating each Condition Occurrence as an independent event.}
\item{It allows aggregation of multiple, closely timed doctor visits for the same Condition to avoid double-counting the Condition Occurrences. For example, consider a Person who visits her Primary Care Physician (PCP) and who is referred to a specialist. At a later time, the Person visits the specialist, who confirms the PCP’s original diagnosis and provides the appropriate treatment to resolve the condition. These two independent doctor visits should be aggregated into one Condition Era.}
}
}

\subsection{ETL Conventions}{

Each Condition Era corresponds to one or many Condition Occurrence records that form a continuous interval. The condition_concept_id field contains Concepts that are identical to those of the CONDITION_OCCURRENCE table records that make up the Condition Era. In contrast to Drug Eras, Condition Eras are not aggregated to contain Conditions of different hierarchical layers. The SQl Script for generating CONDITION_ERA records can be found here The Condition Era Start Date is the start date of the first Condition Occurrence. The Condition Era End Date is the end date of the last Condition Occurrence. Condition Eras are built with a Persistence Window of 30 days, meaning, if no occurrence of the same condition_concept_id happens within 30 days of any one occurrence, it will be considered the condition_era_end_date.
}
}
\keyword{datasets}
