% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\docType{data}
\name{death}
\alias{death}
\title{death}
\format{
A data frame with 4 rows and 7 variables:
\describe{
\item{person_id}{}
\item{death_date}{The date the person was deceased. }
\item{death_datetime}{}
\item{death_type_concept_id}{This is the provenance of the death record, i.e., where it came from. It is possible that an administrative claims database would source death information from a government file so do not assume the Death Type is the same as the Visit Type, etc. }
\item{cause_concept_id}{This is the Standard Concept representing the Person’s cause of death, if available. }
\item{cause_source_value}{}
\item{cause_source_concept_id}{}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#death}
}
\usage{
death
}
\description{
\subsection{Table Description}{

The death domain contains the clinical event for how and when a Person dies. A person can have up to one record if the source system contains evidence about the Death, such as: Condition in an administrative claim, status of enrollment into a health plan, or explicit record in EHR data.
}
}
\keyword{datasets}
